Added exports.quickstart for cfn-response, updated template
Added exports.quickstart to spicecsm.js. Checks for 'Create' resource
type and makes a call to the automated reader on the SpiceCSM corporate
site to check for the subdomain and if it has an automation suite
product.

Updated YAML template to reference the quickstart function,
removed Automated Reader Path (it will be static), changed Spice
Customer Name to SpiceCSM Subdomain.